Worked on High Score
I have the high score variable incrementing and decrementing, with just a little work i'll be able to make it store an organized list of scores. I'll use the overflow property in the text field to try and fit a scrolling 100 top scores chart. Then we just link all of the enemies to the current score variable that runs the script and we're good to go